+++ src/HainanSettlementTool.Core/Services/IHainanEmployeePowerRewardExcelGateway.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using HainanSettlementTool.Core.Models;
+
+namespace HainanSettlementTool.Core.Services
+{
+    public interface IHainanEmployeePowerRewardExcelGateway
+    {
+        IList<HainanEmployeePowerRewardLedgerRow> ReadLedgerRows(HainanEmployeePowerRewardOptions options);
+        HainanEmployeePowerRewardOutput GenerateWorkbooks(HainanEmployeePowerRewardOptions options, HainanEmployeePowerRewardResult result);
+    }
+}